Refactored functionality for template profiles.

+++ Api/Modules/Templates/Models/Preview/PreviewProfileModel.cs
@@ -5,7 +5,7 @@ namespace Api.Modules.Templates.Models.Preview
 {
     public class PreviewProfileModel
     {
-        public Int64 Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Url { get; set; }
         public List<PreviewVariableModel> Variables { get; set; }
